+++ median.h
@@ -1,6 +1,6 @@
 #ifndef MEDIAN_H
 #define MEDIAN_H
 
-float median_find(int* x, int n);
+float median_find(float* x, int n);
 
 #endif
